refactoring KAK_decomposition.py
