--- examples/transpiler/test
@@ -1,3 +1,0 @@
-testfile Changsoo Kim
-
-local test
